# message-mesh-svc/src/main/resources/message-mesh-openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Message Mesh API",
...
    "contact": {
      "name": "Message Mesh Platform Team",
      "url": "https://message-mesh.dev.com",
      "email": "platform@message-mesh.dev.com"
    },
    "license": {
      "name": "Apache License 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.message-mesh.prod.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/messages/{id}/reactions": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Add a reaction",
        "description": "Adds an emoji reaction from the caller (from a fixed palette). Idempotent per user+emoji. The updated message is re-broadcast to the conversation.",
        "operationId": "addReaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Unsupported emoji or not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Reaction added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/messages/{id}": {
      "delete": {
        "tags": [
          "Messages"
        ],
        "summary": "Delete a message",
        "description": "Soft-deletes a message (its body is cleared and it renders as a tombstone). The author or a group admin may delete. The tombstone is re-broadcast to the conversation.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message deleted (tombstone returned)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "400": {
            "description": "Not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Messages"
        ],
        "summary": "Edit a message",
        "description": "Updates the body of a message. Only the original author may edit, and deleted messages cannot be edited. The updated message is re-broadcast to the conversation.",
        "operationId": "edit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Message not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "400": {
            "description": "Not the author, message deleted, or invalid body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Message updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{id}/messages": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "summary": "Fetch message history",
        "description": "Returns messages for a conversation ordered by ascending sequence number. Use 'afterSeq' for forward, cursor-style pagination: pass the highest 'seq' you already hold to fetch only newer messages, and bound the page with 'limit'.",
        "operationId": "messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "afterSeq",
            "in": "query",
            "description": "Return only messages with a sequence greater than this value (cursor)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of messages to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            },
            "example": 50
          }
        ],
        "responses": {
          "403": {
            "description": "Caller is not a member of the conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Messages returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/conversations/{id}/messages": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List messages in a conversation",
        "description": "Returns a page of messages (newest first) for moderation review, including soft-deleted tombstones. 'page' (0-based) and 'size' (max 100) paginate.",
        "operationId": "messages_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Caller is not an administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Messages returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseMessageDto"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/messages/{id}/reactions/{emoji}": {
      "delete": {
        "tags": [
          "Messages"
        ],
        "summary": "Remove a reaction",
        "description": "Removes the caller's emoji reaction. The updated message is re-broadcast to the conversation.",
        "operationId": "removeReaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "emoji",
            "in": "path",
            "description": "URL-encoded emoji to remove",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Reaction removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/conversations/{id}/messages/{messageId}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a message (moderation)",
        "description": "Soft-deletes an individual message within a conversation and broadcasts the tombstone to connected members in real time.",
        "operationId": "deleteMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Message or conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Message deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "ReactionRequest": {
        "required": [
          "emoji"
        ],
        "type": "object",
        "properties": {
          "emoji": {
            "maxLength": 16,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "senderUsername": {
            "type": "string"
          },
          "seq": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SENT",
              "DELIVERED",
              "READ"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "clientTempId": {
            "type": "string"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deleted": {
            "type": "boolean"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "parentPreview": {
            "$ref": "#/components/schemas/ParentPreview"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReactionSummary"
            }
          }
        }
      },
      "ParentPreview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "senderUsername": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "ReactionSummary": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "usernames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "EditMessageRequest": {
        "required": [
          "body"
        ],
        "type": "object",
        "properties": {
          "body": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "PagedResponseMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT bearer token obtained from the authentication endpoints. Paste only the raw token value; Swagger UI adds the 'Bearer ' prefix.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}